pyproject.toml
[project]
name = "pneumonia-classifier-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "scipy>=1.16.3",
    "streamlit>=1.51.0",
    "tensorflow>=2.20.0",
]
